move to existing external test support endpoint
